DOCSP-51123: CC edits and cleanup (#279)
* DOCSP-51123: CC edits and cleanup

* build

* composer

* code errors

* commas

* more phpcs errors

* edits

* phpcs

* code fixes

* more phpcs

* fix

* RR feedback
