+++ composer/composer.json
@@ -2,8 +2,8 @@
   "name": "mongodb/docs-php-library",
   "description": "MongoDB PHP Library Documentation",
   "require": {
-    "ext-mongodb": "^1.20",
-    "mongodb/mongodb": "^1.20"
+    "ext-mongodb": "^2.1",
+    "mongodb/mongodb": "^2.1"
   },
   "require-dev": {
     "doctrine/coding-standard": "^12.0",
