+++ source/data-formats/modeling-bson-data.txt
@@ -57,7 +57,7 @@ these values into BSON documents when inserting them into a collection.
 The following example creates a BSON document that represents the preceding
 :ref:`sample BSON document <php-bson-sample>`:
 
-.. literalinclude:: /includes/bson.php
+.. literalinclude:: /includes/bson/interact-bson.php
     :language: php
     :dedent:
     :start-after: start-create-doc
@@ -73,7 +73,7 @@ to the :ref:`sample BSON document <php-bson-sample>`:
 - Adds a new ``restaurant_id`` field that has a value of ``12345``
 - Changes the ``name`` field value to ``"Galaxy Pizza"``
 
-.. literalinclude:: /includes/bson.php
+.. literalinclude:: /includes/bson/interact-bson.php
     :language: php
     :dedent:
     :start-after: start-modify-doc
@@ -142,7 +142,7 @@ Custom Type Map Example
 The following example sets the ``typeMap`` option for the ``restaurants`` collection
 that serializes arrays and BSON documents as ``MongoDB\Model\BSONDocument`` objects:
 
-.. literalinclude:: /includes/bson.php
+.. literalinclude:: /includes/bson/interact-bson.php
     :language: php
     :dedent:
     :start-after: start-type-map
@@ -170,7 +170,7 @@ Consider the following ``Person`` class definition, which implements the
 ``Persistable`` interface and specifies how to serialize and deserialize
 object fields as BSON values:
 
-.. literalinclude:: /includes/bson.php
+.. literalinclude:: /includes/bson/Person.php
     :language: php
     :dedent:
     :start-after: start-person-class
@@ -182,7 +182,7 @@ database, and reads it back as an object of the same type:
 .. io-code-block::
    :copyable:
 
-   .. input:: /includes/bson.php
+   .. input:: /includes/bson/interact-bson.php
       :start-after: start-person-serialize
       :end-before: end-person-serialize
       :language: php
@@ -245,16 +245,25 @@ and ``bsonUnserialize()`` methods in your class definition.
 Example
 ```````
 
-Consider the following ``User`` class definition, which specifies
-logic for serializing and deserializing its fields into BSON values.
-The class includes a ``role`` field, which has a backed enum value:
+Consider the following backed enum named ``Role``, which has two
+integer-valued cases:
 
-.. literalinclude:: /includes/bson.php
+.. literalinclude:: /includes/bson/Role.php
     :language: php
-    :emphasize-lines: 15, 24
     :dedent:
     :start-after: start-backed-enum
     :end-before: end-backed-enum
+
+This ``User`` class definition includes a ``role`` field with a ``Role`` enum
+value and specifies logic for serializing and deserializing its fields into BSON
+values:
+
+.. literalinclude:: /includes/bson/User.php
+    :language: php
+    :emphasize-lines: 5
+    :dedent:
+    :start-after: start-user-class
+    :end-before: end-user-class
 
 The following example constructs a ``User`` object with a ``role`` field,
 inserts it into the database, and reads it back as an object of the same type:
@@ -262,7 +271,7 @@ inserts it into the database, and reads it back as an object of the same type:
 .. io-code-block::
    :copyable:
 
-   .. input:: /includes/bson.php
+   .. input:: /includes/bson/interact-bson.php
       :start-after: start-enum-serialize
       :end-before: end-enum-serialize
       :language: php
