+++ source/get-started.txt
@@ -186,9 +186,8 @@ The connection string includes the hostname or IP address and
 port of your deployment, the authentication mechanism, user credentials 
 when applicable, and connection options.
 
-.. TODO:
- To connect to an instance or deployment not hosted on Atlas, see
- :ref:`php-connection-targets`.
+To connect to an instance or deployment not hosted on Atlas, see
+:ref:`php-connection-targets`.
 
 .. procedure::
    :style: connected
