+++ source/includes/bson/Role.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Bson;
+
+// start-backed-enum
+enum Role: int
+{
+    case USER = 1;
+    case ADMIN = 2;
+}
+// end-backed-enum
