+++ source/includes/monitoring-logging/logging.php
@@ -3,8 +3,8 @@
 require 'vendor/autoload.php';
 
 // start-monolog-logger
-use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
+use Monolog\Logger;
 
 $logger = new Logger('mongodb-logger');
 $logger->pushHandler(new StreamHandler(__DIR__ . '/mongodb.log', Logger::DEBUG));
@@ -13,16 +13,11 @@ MongoDB\add_logger($logger);
 // end-monolog-logger
 
 // start-custom-logger
-use Psr\Log\AbstractLogger;
-use Psr\Log\LoggerInterface;
-use Psr\Log\LogLevel;
-use MongoDB\PsrLogAdapter;
-
-class MyLogger extends AbstractLogger
+class MyLogger extends Psr\Log\AbstractLogger
 {
     public array $logs = [];
 
-    public function log($level, $message, array $context = []): void
+    public function log(string $level, string|\Stringable $message, array $context = []): void
     {
         $this->logs[] = [$level, $message, $context['domain']];
     }
