+++ source/monitoring-logging/monitoring.txt
@@ -32,9 +32,11 @@ events to track your application's performance and resource use.
 The {+php-library+} emits command events and Server Discovery and Monitoring (SDAM)
 events that provide application information. You can listen for these events to
 monitor your application.
+
+.. note::
    
-.. TODO: This page explains how to monitor your application in code. To learn how to
- record this information to an external log, see the :ref:`php-logging` guide.
+   This page explains how to monitor your application in code. To learn how to
+   record this information to an external log, see the :ref:`php-logging` guide.
 
 Event Types
 -----------
@@ -124,7 +126,7 @@ the ``CommandSubscriber`` interface. The class defines each ``CommandSubscriber`
 and includes custom logic for the ``commandStarted()`` method, which prints details
 about each new command that the server receives:
 
-.. literalinclude:: /includes/monitoring-logging/monitor.php
+.. literalinclude:: /includes/monitoring-logging/MyCommandSubscriber.php
    :start-after: start-command-subscriber
    :end-before: end-command-subscriber
    :language: php
@@ -191,7 +193,7 @@ the ``SDAMSubscriber`` interface. The class defines each ``SDAMSubscriber`` meth
 and includes custom logic for the ``serverOpening()`` method, which prints details
 about servers added to the topology:
 
-.. literalinclude:: /includes/monitoring-logging/monitor.php
+.. literalinclude:: /includes/monitoring-logging/MySdamSubscriber.php
    :start-after: start-sdam-subscriber
    :end-before: end-sdam-subscriber
    :language: php
